Add helper parse function "ordered_dump".
