+++ lib/tower_cli/utils/parser.py
@@ -21,6 +21,7 @@ import shlex
 import sys
 
 from tower_cli.utils import exceptions as exc, debug
+from tower_cli.utils.data_structures import OrderedDict
 
 
 def parse_kv(var_string):
@@ -140,3 +141,26 @@ def process_extra_vars(extra_vars_list, force_json=True):
     if extra_vars == {}:
         return ""
     return json.dumps(extra_vars)
+
+
+def ordered_dump(data, Dumper=yaml.Dumper, **kws):
+    """Expand PyYAML's built-in dumper to support parsing OrderedDict. Return
+    a string as parse result of the original data structure, which includes
+    OrderedDict.
+
+    Args:
+        data: the data structure to be dumped(parsed) which is supposed to
+        contain OrderedDict.
+        Dumper: the yaml serializer to be expanded and used.
+        kws: extra key-value arguments to be passed to yaml.dump.
+    """
+    class OrderedDumper(Dumper):
+        pass
+
+    def _dict_representer(dumper, data):
+        return dumper.represent_mapping(
+            yaml.resolver.BaseResolver.DEFAULT_MAPPING_TAG,
+            data.items())
+    OrderedDumper.add_representer(OrderedDict,
+                                  _dict_representer)
+    return yaml.dump(data, None, OrderedDumper, **kws)
